I don't believe the ":" should be forced on the label tag.
There are cases when the label is used to the right of the form field, and the ":" at the end will look out of place. There are also cases where a "?" would want to be used, such as in cases with "Remember me?" checkboxes. Leave it up to the coder to determine what to use.